add story header
